refactor: extract equations_data string literals from equations_popup
